+++ CHANGELOG.md
@@ -19,6 +19,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## @meteora-ag/alpha-vault [1.1.9] [PR #9](https://github.com/MeteoraAg/alpha-vault-sdk/pull/9)
+
+### Fixed
+
+- Fix bug `getDepositInfo` method
+
 ## @meteora-ag/alpha-vault [1.1.8] [PR #8](https://github.com/MeteoraAg/alpha-vault-sdk/pull/8)
 
 ### Fixed